Extract graphql-backend extentions to separate plugin
