+++ plugins/graphql-backend-node/CHANGELOG.md
@@ -1,0 +1,1 @@
+# @backstage/plugin-graphql-backend-node
