+++ plugins/graphql-backend-node/README.md
@@ -1,0 +1,3 @@
+# plugin-graphql-backend-node
+
+The `graphql-backend-node` plugin contains extension points.
