+++ plugins/graphql-backend-node/api-report.md
@@ -1,0 +1,47 @@
+## API Report File for "@backstage/plugin-graphql-backend-node"
+
+> Do not edit this file. It is a report generated by [API Extractor](https://api-extractor.com/).
+
+```ts
+import { BatchLoadFn } from '@backstage/plugin-graphql-common';
+import { ExtensionPoint } from '@backstage/backend-plugin-api';
+import { GraphQLContext } from '@backstage/plugin-graphql-common';
+import { Module } from 'graphql-modules';
+import { Options } from 'dataloader';
+import { Plugin as Plugin_2 } from 'graphql-yoga';
+
+// @public (undocumented)
+export interface GraphQLApplicationExtensionPoint {
+  // (undocumented)
+  addModule(
+    module: (() => Module | Promise<Module>) | Module | Promise<Module>,
+  ): void;
+  // (undocumented)
+  addSchema(schema: string): void;
+}
+
+// @public (undocumented)
+export const graphqlApplicationExtensionPoint: ExtensionPoint<GraphQLApplicationExtensionPoint>;
+
+// @public (undocumented)
+export interface GraphQLYogaExtensionPoint {
+  // (undocumented)
+  addLoader(name: string, loader: BatchLoadFn<GraphQLContext>): void;
+  // (undocumented)
+  addPlugin(plugin: Plugin_2): void;
+  // (undocumented)
+  setContext<TContext extends Record<string, any>>(
+    context:
+      | ((initialContext: GraphQLContext) => TContext | Promise<TContext>)
+      | Promise<TContext>
+      | TContext,
+  ): void;
+  // (undocumented)
+  setDataloaderOptions(options: Options<string, any>): void;
+}
+
+// @public (undocumented)
+export const graphqlYogaExtensionPoint: ExtensionPoint<GraphQLYogaExtensionPoint>;
+
+// (No @packageDocumentation comment for this package)
+```
