+++ plugins/graphql-backend-node/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "@backstage/plugin-graphql-backend",
-  "description": "An experimental Backstage backend plugin for GraphQL",
-  "version": "0.1.38",
+  "name": "@backstage/plugin-graphql-backend-node",
+  "description": "An experimental Backstage backend extensions plugin for GraphQL",
+  "version": "0.0.0",
   "main": "src/index.ts",
   "types": "src/index.ts",
   "license": "Apache-2.0",
@@ -33,24 +33,14 @@
     "clean": "backstage-cli package clean"
   },
   "dependencies": {
-    "@backstage/backend-common": "workspace:^",
     "@backstage/backend-plugin-api": "workspace:^",
-    "@backstage/plugin-graphql-backend-node": "workspace:^",
     "@backstage/plugin-graphql-common": "workspace:^",
-    "@envelop/core": "^4.0.0",
-    "@envelop/dataloader": "^5.0.0",
-    "@envelop/graphql-modules": "^5.0.0",
     "dataloader": "^2.1.0",
-    "express-promise-router": "^4.1.0",
-    "graphql": "^16.6.0",
     "graphql-modules": "^2.1.0",
-    "graphql-yoga": "^4.0.3",
-    "helmet": "^6.0.0"
+    "graphql-yoga": "^4.0.3"
   },
   "devDependencies": {
-    "@backstage/cli": "workspace:^",
-    "@types/supertest": "^2.0.8",
-    "supertest": "^6.1.3"
+    "@backstage/cli": "workspace:^"
   },
   "files": [
     "dist"
